+++ tatsu/grammars.py
@@ -7,7 +7,7 @@ from collections import defaultdict, Mapping
 from copy import copy
 
 from tatsu.util import indent, trim, ustr, urepr, strtype, compress_seq, chunks
-from tatsu.util import re, RE_FLAGS
+from tatsu.util import re
 from tatsu.exceptions import FailedRef, GrammarError
 from tatsu.ast import AST
 from tatsu.contexts import ParseContext
@@ -281,7 +281,7 @@ class Pattern(Model):
         if not isinstance(ast, list):
             ast = [ast]
         self.patterns = ast
-        re.compile(self.pattern, RE_FLAGS)
+        re.compile(self.pattern)
 
     @property
     def pattern(self):