+++ tatsu/test/grammar/pattern_test.py
@@ -5,6 +5,7 @@ import unittest
 
 from tatsu.util import trim
 from tatsu.tool import compile
+from tatsu.exceptions import FailedParse
 
 
 class PatternTests(unittest.TestCase):
@@ -64,3 +65,41 @@ class PatternTests(unittest.TestCase):
         self.assertEqual(['abc123', 'def456'], ast)
         print(model.pretty())
         self.assertEqual(trim(pretty), model.pretty())
+
+    def test_ignorecase_not_for_pattern(self):
+        grammar = '''
+            @@ignorecase
+            start
+                =
+                {word} $
+                ;
+
+            word
+                =
+                /[a-z]+/
+                ;
+        '''
+
+        model = compile(grammar=grammar)
+        try:
+            model.parse('ABcD xYZ')
+            self.fail('@@ignorecase should not apply to patterns')
+        except FailedParse:
+            pass
+
+    def test_ignorecase_pattern(self):
+        grammar = '''
+            start
+                =
+                {word} $
+                ;
+
+            word
+                =
+                /(?i)[a-z]+/
+                ;
+        '''
+
+        model = compile(grammar=grammar)
+        ast = model.parse('ABcD xYZ')
+        self.assertEqual(['ABcD', 'xYZ'], ast)